fixing the details of the restaurant

+++ Resturaunt/ResturantAPI/ResturantAPI/Models/Menu.cs
@@ -6,6 +6,6 @@ public class Menu
     [Key]
     public int? Id { get; set; }
     public string? Name { get; set; }
-    public double? Price { get; set; }
+    public decimal? Price { get; set; }
     public int? RestId { get; set; }
 }
